+++ Src/Library/changelog.md
@@ -1,4 +1,5 @@
 ### IMPROVEMENTS
-- swagger support for routes params with mixed patterns #380
+- add overload for `TagDescriptions()` method for adding swagger tag descriptions with tuples
+- swagger support for route params with mixed patterns #380
 - misc. improvements to `FastEndpoints.Security` package internals
 - upgrade all dependencies to latest